split the slider widget in an interger and float version
svn path=/trunk/koffice/; revision=1104071

+++ krita/ui/tool/kis_tool_paint.h
@@ -115,7 +115,7 @@ private slots:
 
     void updateCompositeOpComboBox();
     void slotPopupQuickHelp();
-    void slotSetOpacity(qreal opacityPerCent);
+    void slotSetOpacity(int opacityPerCent);
     void slotSetCompositeMode(const QString& compositeOp);
 
 protected slots:
